watch/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Werkzeug==3.0.1
WTForms==3.1.1
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
pytesseract==0.3.10
Flask-Mail==0.9.1
cryptography==42.0.8
psycopg2-binary==2.9.9